Вспомогательные функции перенесены в отдельный модуль
